[CHORE/#56] RetrofitModule 의존성 분리
