+++ app/src/main/java/co/orange/ddanzi/di/module/RetrofitModule.kt
@@ -1,109 +1,19 @@
 package co.orange.ddanzi.di.module
 
-import co.orange.core.extension.isJsonArray
-import co.orange.core.extension.isJsonObject
 import co.orange.ddanzi.BuildConfig.BASE_URL
-import co.orange.ddanzi.di.interceptor.AuthInterceptor
-import co.orange.ddanzi.di.interceptor.DeviceTokenInterceptor
-import co.orange.ddanzi.di.qualifier.RetrofitQualifier
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import co.orange.ddanzi.di.interceptor.RetrofitQualifier
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import kotlinx.serialization.json.Json
-import okhttp3.Interceptor
-import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import org.json.JSONArray
-import org.json.JSONObject
 import retrofit2.Converter
 import retrofit2.Retrofit
-import timber.log.Timber
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object RetrofitModule {
-    private const val APPLICATION_JSON = "application/json"
-
-    @Provides
-    @Singleton
-    fun provideJson(): Json =
-        Json {
-            ignoreUnknownKeys = true
-            prettyPrint = true
-        }
-
-    @Provides
-    @Singleton
-    fun provideJsonConverter(json: Json): Converter.Factory =
-        json.asConverterFactory(
-            APPLICATION_JSON.toMediaType(),
-        )
-
-    @Provides
-    @Singleton
-    fun provideHttpLoggingInterceptor(): Interceptor =
-        HttpLoggingInterceptor { message ->
-            when {
-                message.isJsonObject() ->
-                    Timber.tag("okhttp").d(JSONObject(message).toString(4))
-
-                message.isJsonArray() ->
-                    Timber.tag("okhttp").d(JSONArray(message).toString(4))
-
-                else -> {
-                    Timber.tag("okhttp").d("CONNECTION INFO -> $message")
-                }
-            }
-        }.apply {
-            level = HttpLoggingInterceptor.Level.BODY
-        }
-
-    @Provides
-    @Singleton
-    @RetrofitQualifier.JWT
-    fun provideAuthInterceptor(authInterceptor: AuthInterceptor): Interceptor = authInterceptor
-
-    @Provides
-    @Singleton
-    @RetrofitQualifier.DEVICE
-    fun provideDeviceTokenInterceptor(deviceTokenInterceptor: DeviceTokenInterceptor): Interceptor = deviceTokenInterceptor
-
-    @Provides
-    @Singleton
-    @RetrofitQualifier.JWT
-    fun provideJWTOkHttpClient(
-        loggingInterceptor: Interceptor,
-        @RetrofitQualifier.JWT authInterceptor: Interceptor,
-    ): OkHttpClient =
-        OkHttpClient.Builder()
-            .addInterceptor(loggingInterceptor)
-            .addInterceptor(authInterceptor)
-            .build()
-
-    @Provides
-    @Singleton
-    @RetrofitQualifier.NOTOKEN
-    fun provideReissueOkHttpClient(loggingInterceptor: Interceptor): OkHttpClient =
-        OkHttpClient.Builder()
-            .addInterceptor(loggingInterceptor)
-            .build()
-
-    @Provides
-    @Singleton
-    @RetrofitQualifier.DEVICE
-    fun provideDeviceOkHttpClient(
-        loggingInterceptor: Interceptor,
-        @RetrofitQualifier.DEVICE deviceTokenInterceptor: Interceptor,
-    ): OkHttpClient =
-        OkHttpClient.Builder()
-            .addInterceptor(loggingInterceptor)
-            .addInterceptor(deviceTokenInterceptor)
-            .build()
-
     @Provides
     @Singleton
     @RetrofitQualifier.JWT
